Rename polymer/element-descriptor.ts to polymer-element.ts
Move toScannedPolymerProperty() into polymer/
